--- MusicCatalog/src/utils/checkPrice.js
@@ -1,1 +1,0 @@
-export const checkPrice = price => (price>0 ? price + ' $' : 'FREE');
