--- MusicCatalog/src/utils/timeConventer.js
@@ -1,5 +1,0 @@
-export const timeConventer = millis => {
-  let minutes = Math.floor(millis / 60000);
-  let seconds = ((millis % 60000) / 1000).toFixed(0);
-  return minutes + ':' + (seconds < 10 ? '0' : '') + seconds;
-};
